feat : 제안서 페이지 path 수정
