+++ src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import LoadingPage from "./pages/LoadingPage";
 import LoginHome from "./pages/LoginHome";
 import MainHome from "./pages/MainHome";
-import Proposal from "./pages/Proposal";
+import CreateProposal from "./pages/CreateProposal";
 import ShopDetailPage from "./pages/ShopDetailPage";
 import AddSchedulePage from "./pages/AddSchedulePage";
 import SearchPage from "./pages/SearchPage";
@@ -11,6 +11,7 @@ import Recommend from "./pages/Recommend";
 import OurService from "./pages/OurService";
 import ShopMemberMyPage from "./pages/ShopMemberMyPage";
 import AllProposal from "./pages/AllProposal";
+import Proposal from "./pages/Proposal";
 
 function App() {
   return (
@@ -24,7 +25,8 @@ function App() {
         <Route exact path="/AddSchedulePage" element={<AddSchedulePage />} />
         <Route exact path="/search" element={<SearchPage />} />
         <Route exact path="/recommend" element={<Recommend />} />
-        <Route path="/proposal/*" element={<Proposal />} />
+        <Route path="/create/*" element={<CreateProposal />} />
+        <Route exact path="/proposal/*" element={<Proposal />} />
         <Route exact path="/serviceinfo" element={<OurService />} />
         <Route exact path="/mypagehost" element={<ShopMemberMyPage />} />
         <Route exact path="/allproposal" element={<AllProposal />} />
