+++ src/pages/CreateProposal/Price.js
@@ -5,8 +5,8 @@ import { Link } from "react-router-dom";
 import SmallPinkButton from "../../components/SmallPinkButton";
 import SmallWhiteButton from "../../components/SmallWhiteButton";
 import SmallGrayButton from "../../components/SmallGrayButton";
-import PageTitle from "../../components/PageTitle";
 
+import PageTitle from "../../components/PageTitle";
 import ProposalText from "../../components/Proposal/ProposalText";
 import ProgessBar from "../../components/Proposal/ProgressBar";
 
@@ -38,7 +38,7 @@ const Circle = styled.div`
 
 const Option = styled.p`
   margin: auto 0 auto 23px;
-  width: 75px;
+  width: 150px;
   font-family: "Apple SD Gothic Neo";
   font-style: normal;
   font-weight: 400;
@@ -49,120 +49,65 @@ const Option = styled.p`
   color: var(--black);
 `;
 
-const Exp = styled.p`
-  font-family: "Apple SD Gothic Neo";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 32px;
-  /* identical to box height */
-
-  color: var(--sub-darkgray);
-`;
-
-const DesignInput = styled.div`
-  min-height: 50px;
-  height: auto;
-
-  margin: 38px 24px 0 24px;
-  background: var(--sub-lightgray);
-  border-radius: 6px;
-  border: none;
-  padding: auto;
-
-  textarea {
-    width: 346px;
-    height: 17px;
-
-    resize: none;
-    overflow: auto;
-
-    padding: 17px;
-
-    background: transparent;
-    border: none;
-    color: var(--sub-darkgray);
-
-    font-family: "Apple SD Gothic Neo";
-    font-style: normal;
-    font-weight: 500;
-    font-size: 14px;
-    line-height: 17px;
-  }
-
-  textarea::placeholder {
-    color: var(--background);
-  }
-  textarea:focus {
-    outline: none;
-  }
-`;
-
-const Taste = ({ history, setHistory }) => {
+const Price = ({ history, setHistory }) => {
   const [isChecked, setIsChecked] = useState(null);
 
-  const [tastes, setTastes] = useState([
-    { id: 1, taste: "초코" },
-    { id: 2, taste: "바닐라" },
-    { id: 3, taste: "기타" },
+  const [prices, setPrices] = useState([
+    { id: 1, price: "1만원 미만" },
+    { id: 2, price: "1만원 이상 3만원 미만" },
+    { id: 3, price: "3만원 이상 7만원 미만" },
+    { id: 4, price: "7만원 이상 10만원 미만" },
+    { id: 5, price: "10만원 이상" },
   ]);
 
   const onToggle = id => {
     setIsChecked(id);
   };
 
-  const [isDone, setIsDone] = useState(true);
-
-  const ThisStep = 60;
+  const ThisStep = 75;
 
   return (
     <div>
       <PageTitle title="제안서 작성하기" margin="56px auto 0 auto" />
       <ProgessBar step={ThisStep} before={history} />
 
-      <ProposalText text="어떤 맛을 원하시나요?" />
+      <ProposalText text="원하는 가격대를 선택해주세요." />
 
       <Wrapper>
-        {tastes.map(taste => {
-          if (taste.id === isChecked) {
+        {prices.map(price => {
+          if (price.id === isChecked) {
             return (
               <div
-                onClick={() => onToggle(taste.id)}
+                onClick={() => onToggle(price.id)}
                 className="item"
-                key={taste.id}
+                key={price.id}
                 style={{
                   display: "flex",
                   height: "60px",
                 }}
               >
                 <Circle checked />
-                <Option>{taste.taste}</Option>
+                <Option>{price.price}</Option>
               </div>
             );
           } else {
             return (
               <div
-                onClick={() => onToggle(taste.id)}
+                onClick={() => onToggle(price.id)}
                 className="item"
-                key={taste.id}
+                key={price.id}
                 style={{
                   display: "flex",
                   height: "60px",
                 }}
               >
                 <Circle />
-                <Option>{taste.taste}</Option>
+                <Option>{price.price}</Option>
               </div>
             );
           }
         })}
       </Wrapper>
-
-      <ProposalText text="디자인에 대해 상세히 적어주세요!"></ProposalText>
-
-      <DesignInput>
-        <textarea placeholder="ex) 레터링을 ‘생일 축하해'로 해주세요!" />
-      </DesignInput>
 
       <div
         style={{
@@ -172,16 +117,15 @@ const Taste = ({ history, setHistory }) => {
           justifyContent: "center",
         }}
       >
-        <Link to="/proposal/size">
+        <Link to="/create/taste">
           <SmallWhiteButton onClick={() => setHistory(ThisStep)}>
-            {" "}
             이전
           </SmallWhiteButton>
         </Link>
 
         <div style={{ marginLeft: "6px" }}>
           {isChecked ? (
-            <Link to="/proposal/price">
+            <Link to="/create/design">
               <SmallPinkButton onClick={() => setHistory(ThisStep)}>
                 완료
               </SmallPinkButton>
@@ -195,4 +139,4 @@ const Taste = ({ history, setHistory }) => {
   );
 };
 
-export default Taste;
+export default Price;
